feat: 주문 상세 내역 조회(방장) api

+++ src/main/java/com/ordertogether/team14_be/order/details/dto/get/MemberBriefInfo.java
@@ -1,3 +1,3 @@
 package com.ordertogether.team14_be.order.details.dto.get;
 
-public record MemberBriefInfo(int memberId, String deliveryName, int price, boolean isPayed) {}
+public record MemberBriefInfo(long memberId, String deliveryName, int price, boolean isPayed) {}
